backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-multipart==0.0.6
pdfminer.six==20221105
docx2txt==0.8
celery==5.3.1
redis==4.5.5
beautifulsoup4==4.12.2
chromadb==0.3.26
flask==2.3.2
python-dotenv==1.0.0
